requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Paginate==0.5.4
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2